chore(prisma): rename cli types stub helper
