+++ packages/cli/helpers/write-cli-types-stub.ts
